Add a drag and drop demo
This may expose a bug in Avalonia.FuncUI.  As soon as you press the
mouse button on Drag Me you get two PressedPointerEvents.  This makes
certain cases not work properly. For example,

1. When you immediately release the button you do not get a
   DragDropEffects.None cancellation response.

2. If you drag it to the drop target, you get a count of 1 the first
   time and n + 2 for each subsequent attempt.

If you look at the output in the terminal you can see the multiple
messages being triggered and how the display and state do not match.