Fix onPointerPressed handling
Marks the event as handled avoiding the ugly hack of tracking drag state
in the model.

The remaining bug is that occassionally we get a Dragged message without
receiving a Dropped message. The Dragged message is fired by completion
of the doDrag Async. I assume that is why we always get one. However the
Dropped message is dispatched from the drop target's onDrop handler. It
is possible that this is responsible for the difference.

I tried adding diagnostics to the onDrop handler to see if it was
receiving and unexpected data type that is not being handled, but it is
not even being called.